% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{absolute.observed.benefit}
\alias{absolute.observed.benefit}
\title{calculates the absolute observed benefit, which corresponds to the difference in \code{mean(y[W=w])}, and the associated confidence interval.}
\usage{
absolute.observed.benefit(y, w, significance.level = 0.05)
}
\arguments{
\item{y}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
calculates the absolute observed benefit, which corresponds to the difference in \code{mean(y[W=w])}, and the associated confidence interval.
}
