% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{average_treatment_effect}
\alias{average_treatment_effect}
\title{computes the average treatment effect (ATE) of a predmod object}
\usage{
average_treatment_effect(
  x,
  subset = NULL,
  relative = FALSE,
  neww = NULL,
  newX = NULL,
  newz = NULL,
  shrunk = FALSE
)
}
\arguments{
\item{x}{A predmod object}

\item{subset}{The indices of the subgroup of interest}

\item{relative}{Shall relative ATE be calculated?}

\item{neww}{Optional treatment assignment variables to calculate benefits with}

\item{newX}{Optional covariate matrix to calculate benefits with}

\item{newz}{Optional linear predictors}

\item{shrunk}{TODO}
}
\description{
computes the average treatment effect (ATE) of a predmod object
}
