% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-risk.R
\name{baseline.risk}
\alias{baseline.risk}
\title{baseline risk prediction via penalized logistic regression (treatment assignment is not used)}
\usage{
baseline.risk(X, y, alpha = 1)
}
\arguments{
\item{X}{matrix of data frame of covariates}

\item{y}{vector of binary responses}

\item{alpha}{the alpha as in glmnet. Defauly is 1 (= Lasso). If NULL, no penalty is applied.}
}
\description{
baseline risk prediction via penalized logistic regression (treatment assignment is not used)
}
