% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-risk.R
\name{baseline.risk.survival}
\alias{baseline.risk.survival}
\title{baseline survival}
\usage{
baseline.risk.survival(X, status, time, alpha = NULL)
}
\arguments{
\item{X}{covariate matrix}

\item{status}{binary status. 1 if failure, 0 if survived}

\item{time}{right-censored time at risk}

\item{alpha}{penalty}
}
\description{
baseline survival
}
