% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-riskNEW.R
\name{baseline_survival}
\alias{baseline_survival}
\title{baseline risk prediction via penalized logistic regression (treatment assignment is not used)}
\usage{
baseline_survival(
  X,
  status,
  time,
  time_eval = max(time),
  alpha = 1,
  failcode = 1,
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{status}{Vector of binary responses}

\item{time}{Time at risk}

\item{time_eval}{timat at which survival shall be evaluated}

\item{alpha}{The elasticnet mixing parameter for regularization, with \eqn{0\le\alpha\le 1}.
The penalty is defined as \deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} (default) is the lasso penalty, and \code{alpha=0} the ridge penalty. If \code{NULL}, no regularization is used.}

\item{failcode}{Code of status that denotes the failure type of interest. The code 0 indicates censoring.}

\item{...}{Additional arguments to be passed}
}
\description{
baseline risk prediction via penalized logistic regression (treatment assignment is not used)
}
