% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-statistics.R
\name{C.index.benefit}
\alias{C.index.benefit}
\title{calculates C index for benefit}
\usage{
C.index.benefit(y, w, predicted.benefit)
}
\arguments{
\item{y}{binary response vector}

\item{w}{binary treatment assignment vector}

\item{predicted.benefit}{vector of predicted benefits}
}
\value{
point estimate and standard error of C index for benefit
}
\description{
calculates C index for benefit
}
