% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-statistics.R
\name{C.index.outcome}
\alias{C.index.outcome}
\title{calculates C index outcome}
\usage{
C.index.outcome(y, risk.prediction)
}
\arguments{
\item{y}{binary response vector}

\item{risk.prediction}{vector of risk predictions (i.e. Pr(y = 1))}
}
\value{
point estimate and standard error of C index outcome
}
\description{
calculates C index outcome
}
