% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{C_outcome}
\alias{C_outcome}
\title{calculates C index outcome}
\usage{
C_outcome(y, risk)
}
\arguments{
\item{y}{binary response vector}

\item{risk}{vector of risk predictions (i.e. Pr(y = 1))}
}
\value{
point estimate and standard error of C index outcome
}
\description{
calculates C index outcome
}
