% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calibration.plot.grf}
\alias{calibration.plot.grf}
\title{makes a calibration plot for a GRF model}
\usage{
calibration.plot.grf(
  grf.model.obj,
  cutoffs = c(0.25, 0.5, 0.75),
  risk.baseline = NULL,
  significance.level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip.sign.of.absolute.benefit = FALSE
)
}
\arguments{
\item{grf.model.obj}{GRF model oobject as returned by grf.modeling()}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is \code{c(0.25, 0.5, 0.75)}, which corresponds to the quartiles.}

\item{risk.baseline}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{pred.model.obj} is used.}

\item{significance.level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-xcis}

\item{flip.sign.of.absolute.benefit}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes a calibration plot for a GRF model
}
