% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calibration.plot_imputation.accounter}
\alias{calibration.plot_imputation.accounter}
\title{makes a calibration plot for a prediction model while accounting for imputation uncertainty}
\usage{
calibration.plot_imputation.accounter(
  pred.model.objs_imputed,
  cutoffs = c(0.25, 0.5, 0.75),
  relative = FALSE,
  significance.level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip.sign.of.absolute.benefit = FALSE
)
}
\arguments{
\item{pred.model.objs_imputed}{prediction model object, as returned by risk.modeling() or effect.modeling() TODO: use name of imputation functions}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is c(0.25, 0.5, 0.75), which corresponds to the quartiles.}

\item{relative}{logical. If TRUE, then relative benefits will be plotted. Default is FALSE}

\item{significance.level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-xcis}

\item{flip.sign.of.absolute.benefit}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes a calibration plot for a prediction model while accounting for imputation uncertainty
}
