% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{makes a calibration plot for a prediction model}
\usage{
calibration_plot(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  relative = FALSE,
  baseline_risk = NULL,
  time_eval = NULL,
  significance_level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip_sign = FALSE
)
}
\arguments{
\item{x}{prediction model object, as returned by risk.modeling() or effect.modeling()}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is \code{c(0.25, 0.5, 0.75)}, which corresponds to the quartiles.}

\item{relative}{logical. If \code{TRUE}, then relative benefits will be plotted. Default is \code{FALSE}}

\item{baseline_risk}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{x} is used.}

\item{time_eval}{Time at which we evaluate the risk predictions.}

\item{significance_level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-xcis}

\item{flip_sign}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes a calibration plot for a prediction model
}
