% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{calibration_plot_grf}
\alias{calibration_plot_grf}
\title{makes a calibration plot for a prediction model}
\usage{
calibration_plot_grf(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  baseline_risk = NULL,
  significance_level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip_sign = FALSE
)
}
\arguments{
\item{x}{grf model}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is \code{c(0.25, 0.5, 0.75)}, which corresponds to the quartiles.}

\item{baseline_risk}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk}

\item{significance_level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-xcis}

\item{flip_sign}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes a calibration plot for a prediction model
}
