% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-impaccount.R
\name{calibration_plot_grf_imputation}
\alias{calibration_plot_grf_imputation}
\title{makes a calibration plot for a GRF model and adjusts for imputation uncertainty}
\usage{
calibration_plot_grf_imputation(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  baseline_risk = NULL,
  significance_level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip_sign = FALSE
)
}
\arguments{
\item{x}{A list of grf models}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is \code{c(0.25, 0.5, 0.75)}, which corresponds to the quartiles.}

\item{baseline_risk}{A list of baseline risks that shall be used for grouping. If \code{NULL} (default), then the baseline risks in \code{x} are used.}

\item{significance_level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-axis}

\item{flip_sign}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes a calibration plot for a GRF model and adjusts for imputation uncertainty
}
