% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits_impaccountMERGE.R
\name{calibration_plot_imputation}
\alias{calibration_plot_imputation}
\title{makes an imputation-accounted calibration plot for imputed prediction models}
\usage{
calibration_plot_imputation(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  relative = FALSE,
  baseline_risk = NULL,
  benefits_risk = FALSE,
  time_eval = NULL,
  significance_level = 0.05,
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  flip_sign = FALSE
)
}
\arguments{
\item{x}{A list of prediction model objects}

\item{cutoffs}{the cutoff points of quantiles that shall be used for GATES grouping. Default is \code{c(0.25, 0.5, 0.75)}, which corresponds to the quartiles.}

\item{relative}{logical. If \code{TRUE}, then relative benefits will be plotted. Default is \code{FALSE}}

\item{baseline_risk}{A list of baseline risks that shall be used for grouping. If \code{NULL} (default), then the baseline risks as in \code{x} are used.}

\item{benefits_risk}{Logical. If \code{TRUE}, then the risk-based benefits are used (only applicable to survival models). Default is \code{FALSE}.}

\item{time_eval}{Only applicable if \code{benefits_risk = TRUE}. Time at which we evaluate the risk predictions.}

\item{significance_level}{significance level for the confidence intervals. Default is 0.05}

\item{title}{optional title of the plot}

\item{xlim}{limits of x-axis}

\item{ylim}{limits of y-xcis}

\item{flip_sign}{logical. Shall the sign of the benefits be flipped?}
}
\description{
makes an imputation-accounted calibration plot for imputed prediction models
}
