% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordance}
\alias{concordance}
\title{Function to predict concordance}
\usage{
concordance(
  x,
  baseline_risk = NULL,
  risk = NULL,
  benefit = NULL,
  status = NULL,
  w = NULL
)
}
\arguments{
\item{x}{A predmod object. TODO}
}
\description{
Function to predict concordance
}
