% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{Performs DML estimation}
\usage{
dml(
  X,
  w,
  y,
  ml_g = "random.forest",
  ml_m = "random.forest",
  significance.level = 0.05,
  n_folds = 5,
  score = "ATE",
  trimming_rule = "truncate",
  trimming_threshold = 1e-12,
  dml_procedure = "dml2",
  draw_sample_splitting = TRUE,
  apply_cross_fitting = TRUE
)
}
\arguments{
\item{X}{a matrix or data frame of covariates}

\item{w}{a binary vector of treatment status}

\item{y}{a binary vector of outcomes}

\item{ml_g}{a regression machine learner; refers to the nuisance function \code{g0(X) = E[Y|X,W]}. Either  'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_g = mlr3::lrn("regr.ranger", num.trees = 500, mtry = NULL, min.node.size = NULL, max.depth = NULL) for a classification forest, which is also the default.}

\item{ml_m}{a classification machine learner; refers to the nuisance function \code{m0(X) = E[W|X]}. Either  'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_m = mlr3::lrn("classif.ranger", num.trees = 500, mtry = NULL, min.node.size = NULL, max.depth = NULL) for a classification forest, which is also the default.}

\item{significance.level}{TODO}

\item{n_folds}{TODO}

\item{score}{TODO}

\item{trimming_rule}{TODO}

\item{trimming_threshold}{TODO}

\item{dml_procedure}{TODO}

\item{draw_sample_splitting}{TODO}

\item{apply_cross_fitting}{TODO}
}
\value{
Estimates of the causal parameter and a 'DoubleML' object
}
\description{
Performs DML estimation
}
