% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-modelingOLD.R
\name{effect.modeling}
\alias{effect.modeling}
\title{performs effect modeling by penalized logistic regression}
\usage{
effect.modeling(
  X,
  y,
  w,
  alpha = 1,
  lifeyears = NULL,
  prediction.timeframe = NULL,
  interacted.variables = NULL,
  retained = FALSE,
  risk.baseline = NULL,
  retain_w = TRUE
)
}
\arguments{
\item{X}{design matrix or data frame}

\item{y}{a binary response vector}

\item{w}{a binary treatment assignment vector}

\item{alpha}{the alpha as in glmnet. Default is 1 (= Lasso)}

\item{lifeyears}{vector of life years. Default is NULL.}

\item{prediction.timeframe}{vector of the prediction time frame. Default is NULL.}

\item{interacted.variables}{character vector of variables that are to be interacted with \code{w}}

\item{retained}{character vector of variables that shall not be regularized}

\item{risk.baseline}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{pred.model.obj} is used.}

\item{retain_w}{Logical. Shall \code{w} be retained?}
}
\description{
TODO: make argument to pass baseline risk as argument
TODO: pass all arguments of baselinerisk() and allow for different alpha
}
