% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_model_crss.R
\name{effect_model}
\alias{effect_model}
\title{Effect model}
\usage{
effect_model(
  X,
  status,
  w,
  interacted = 1:ncol(X),
  alpha = 1,
  failcode = 1,
  retained = NULL,
  retain_w = TRUE,
  baseline_risk = NULL,
  alpha_baseline = 1,
  ...
)
}
\arguments{
\item{X}{Matrix of fixed covariates.}

\item{status}{Numeric vector with a unique code for each failure type. Code 0 denotes a censored observation.}

\item{w}{Binary vector of treatment assignment status. Equal to 1 for treatment group and 0 for control group.}

\item{interacted}{Variables that are to be interacted with \code{w}. Either numeric or character.}

\item{alpha}{The elasticnet mixing parameter for regularization in the effect model. The penalty is defined as \deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} (default) is the lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{failcode}{Code of status that denotes the failure type of interest. Default is one.}

\item{retained}{Character vector of variables that shall not be regularized. Start names with \code{w.} to indicate interacted variables.}

\item{retain_w}{Logical. Shall treatment assignment \code{w} be retained? Defalt is \code{TRUE}.}

\item{baseline_risk}{User-defined baseline risk. If \code{NULL} (default), then baseline risk if estimated with \code{(X, status)} using \code{\link{baseline_risk}}.}

\item{alpha_baseline}{The elasticnet mixing parameter for regularization in a potential baseline risk model to obtain baseline risk. Only applicable if \code{z = NULL}. See \code{\link{baseline_risk}} for details.}

\item{...}{Additional arguments to be passed.}
}
\value{
A \code{predmod_ordinary} object.
}
\description{
Effect model
}
