% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failure_prediction.R
\name{failure}
\alias{failure}
\title{Function to predict the exact failure times}
\usage{
failure(x, time = NULL)
}
\arguments{
\item{x}{A predmod object. TODO}

\item{time}{A vector of times along which to compute the survival curves}
}
\description{
Function to predict the exact failure times
}
