% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{get.benefits}
\alias{get.benefits}
\title{Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get.benefits(
  pred.model.obj,
  cutoffs = c(0.25, 0.5, 0.75),
  significance.level = 0.05
)
}
\arguments{
\item{pred.model.obj}{prediction model object}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
