% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{get.benefits_imputation.accounter}
\alias{get.benefits_imputation.accounter}
\title{Calculates imputation-uncertainty-accounted group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get.benefits_imputation.accounter(
  pred.model.objs_imputed,
  cutoffs = c(0.25, 0.5, 0.75),
  significance.level = 0.05
)
}
\arguments{
\item{pred.model.objs_imputed}{list of prediction model objects}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
Calculates imputation-uncertainty-accounted group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
