% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{get_benefits}
\alias{get_benefits}
\title{Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get_benefits(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  baseline_risk = NULL,
  significance_level = 0.05
)
}
\arguments{
\item{x}{prediction model object}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{baseline_risk}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{pred.model.obj} is used.}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
