% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{get_benefits_grf}
\alias{get_benefits_grf}
\title{Calculates group-level benefits from a GRF model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get_benefits_grf(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  breaks = NULL,
  baseline_risk = NULL,
  significance_level = 0.05
)
}
\arguments{
\item{x}{GRF model object}

\item{cutoffs}{the quantile cutoff points. Default is \code{c(0.25, 0.5, 0.75)}, which yields the quartiles.}

\item{breaks}{Breaks along which to perform the grouping. If passed, overrules the grouping implied by \code{cutoffs}}

\item{baseline_risk}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{x} is used.}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
Calculates group-level benefits from a GRF model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
