% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits-impaccount.R
\name{get_benefits_grf_imputation}
\alias{get_benefits_grf_imputation}
\title{Calculates imputation-adjusted group-level benefits from a GRF model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get_benefits_grf_imputation(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  breaks = NULL,
  baseline_risk = NULL,
  significance_level = 0.05
)
}
\arguments{
\item{x}{GRF model object}

\item{cutoffs}{the quantile cutoff points. Default is \code{c(0.25, 0.5, 0.75)}, which yields the quartiles.}

\item{breaks}{List of breaks along which to perform the grouping. If passed, overrules the grouping implied by \code{cutoffs}}

\item{baseline_risk}{A list of baseline risks that shall be used for grouping. If \code{NULL} (default), then the baseline risks in \code{x} are used.}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
Calculates imputation-adjusted group-level benefits from a GRF model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
