% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{get_benefits_imputation}
\alias{get_benefits_imputation}
\title{Calculates imputation-adjusted group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get_benefits_imputation(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  baseline_risk = NULL,
  time_eval = NULL,
  significance_level = 0.05,
  status = NULL,
  w = NULL,
  X = NULL
)
}
\arguments{
\item{x}{a list of prediction model object}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{baseline_risk}{A list of baseline risks that shall be used for grouping. If \code{NULL} (default), then the baseline risks in \code{x} are used.}

\item{time_eval}{Time at which we evaluate the risk predictions.}

\item{significance_level}{the significance level. Default is 0.05.}

\item{status}{Optional list of target variables to calculate benefits with}

\item{w}{Optional list of treatment assignment variables to calculate benefits with}
}
\description{
Calculates imputation-adjusted group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
