% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf.R
\name{grf.modeling}
\alias{grf.modeling}
\title{applies GRF modeling. Note that risk modeling is not possible here}
\usage{
grf.modeling(
  X,
  y,
  w,
  lifeyears = NULL,
  prediction.timeframe = NULL,
  store.causal.forest.object = TRUE,
  num.trees = 2000,
  ...
)
}
\arguments{
\item{X}{design matrix, can also be a data frame}

\item{y}{vector of binary responses.}

\item{w}{vector of binary treatment assignments}

\item{lifeyears}{vector of life years. Default is NULL.}

\item{prediction.timeframe}{vector of the prediction time frame. Default is NULL.}

\item{store.causal.forest.object}{TODO}

\item{num.trees}{number of trees}

\item{...}{Add arguments}
}
\description{
applies GRF modeling. Note that risk modeling is not possible here
}
