% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf.R
\name{grf_model}
\alias{grf_model}
\title{GRF modeling}
\usage{
grf_model(X, status, w, failcode = 1, num_trees = 2000, ...)
}
\arguments{
\item{X}{Matrix of fixed covariates.}

\item{status}{Numeric vector with a unique code for each failure type. Code 0 denotes a censored observation.}

\item{w}{Binary vector of treatment assignment status. Equal to 1 for treatment group and 0 for control group.}

\item{failcode}{Code of status that denotes the failure type of interest. Default is one.}

\item{num_trees}{number of trees}

\item{...}{Additional arguments}
}
\value{
A list with predicted absolute benefits, the baseline risk, the causal forest, failure type and failure status
}
\description{
Estimates a Generalized Random Forest/Causal Forest as described in Wager&Athey (2018)
}
\references{
Wager S. &  Athey S. (2018) Estimation and Inference of Heterogeneous Treatment Effects using Random Forests,
Journal of the American Statistical Association, 113:523, 1228-1242, DOI: 10.1080/01621459.2017.1319839
}
