% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf.R
\name{grf_model_survival}
\alias{grf_model_survival}
\title{GRF modeling}
\usage{
grf_model_survival(X, status, time, w, failcode = 1, num_trees = 2000, ...)
}
\arguments{
\item{X}{Matrix of fixed covariates.}

\item{status}{Numeric vector with a unique code for each failure type. Code 0 denotes a censored observation.}

\item{time}{Vector of failure/censoring times.}

\item{w}{Binary vector of treatment assignment status. Equal to 1 for treatment group and 0 for control group.}

\item{failcode}{Code of status that denotes the failure type of interest. Default is one.}

\item{num_trees}{number of trees}

\item{...}{Additional arguments}
}
\description{
GRF modeling
}
