% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impaccount}
\alias{impaccount}
\title{Account for imputation uncertainty in predmod models}
\usage{
impaccount(x)
}
\arguments{
\item{x}{a list of predmod model objects (imputed)}
}
\value{
A list of imputation-accounted stuff
}
\description{
Account for imputation uncertainty in predmod models
}
