% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-risk-impaccount.R
\name{impaccount_baseline_risk}
\alias{impaccount_baseline_risk}
\title{Account for imputation uncertainty in baseline risk models}
\usage{
impaccount_baseline_risk(x)
}
\arguments{
\item{x}{a list of baseline risk model objects (imputed)}
}
\value{
A list of imputation-accounted stuff
}
\description{
Account for imputation uncertainty in baseline risk models
}
