% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impaccount_concordance}
\alias{impaccount_concordance}
\title{Account for imputation uncertainty in concordance estimates}
\usage{
impaccount_concordance(x)
}
\arguments{
\item{x}{A list of objects of type \code{"concordance"}}
}
\value{
Imputation accounted concordance estimates
}
\description{
Account for imputation uncertainty in concordance estimates
}
