% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-model-impaccount.R
\name{impaccount_effect_model}
\alias{impaccount_effect_model}
\title{Account for imputation uncertainty in effect models}
\usage{
impaccount_effect_model(x)
}
\arguments{
\item{x}{a list of effect model objects (imputed)}
}
\value{
A list of effect model outcomes (model coefficients, baseline risk, regular risk estimates, counterfactual risk estimates, absolute benefits, relative benefits) which account for imputation uncertainty
}
\description{
Account for imputation uncertainty in effect models
}
