% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{multiple.imputation}
\alias{multiple.imputation}
\title{impute datasets via kNN imputation}
\usage{
multiple.imputation(
  X,
  y,
  w,
  lifeyears = NULL,
  prediction.timeframe = NULL,
  m = 10,
  k = 5
)
}
\arguments{
\item{X}{matrix of covariates}

\item{y}{binary outcome variable}

\item{w}{binary treatment assignment}

\item{lifeyears}{vector of life years. Default is NULL.}

\item{prediction.timeframe}{vector of the prediction time frame. Default is NULL.}

\item{m}{number of imputations}

\item{k}{number of nearest neighbors for donor imputation}
}
\description{
impute datasets via kNN imputation
}
