% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{observed_benefit_absolute}
\alias{observed_benefit_absolute}
\title{calculates the absolute observed benefit, which corresponds to the difference in \code{mean(y[W=w])}, and the associated confidence interval.}
\usage{
observed_benefit_absolute(status, w, significance_level = 0.05)
}
\arguments{
\item{status}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
calculates the absolute observed benefit, which corresponds to the difference in \code{mean(y[W=w])}, and the associated confidence interval.
}
