% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{observed_benefit_relative}
\alias{observed_benefit_relative}
\title{calculates the relative observed benefit, which corresponds to a risk ratio, and the associated confidence interval}
\usage{
observed_benefit_relative(status, w, significance_level = 0.05)
}
\arguments{
\item{status}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
calculates the relative observed benefit, which corresponds to a risk ratio, and the associated confidence interval
}
