% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{odds.ratio}
\alias{odds.ratio}
\title{calculates the odds ratio, and the associated confidence interval.}
\usage{
odds.ratio(y, w, significance.level = 0.05)
}
\arguments{
\item{y}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
calculates the odds ratio, and the associated confidence interval.
}
