% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{odds_ratio}
\alias{odds_ratio}
\title{calculates the odds ratio, and the associated confidence interval.}
\usage{
odds_ratio(status, w, significance_level = 0.05)
}
\arguments{
\item{status}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
calculates the odds ratio, and the associated confidence interval.
}
