% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-risk.R
\name{predict.baseline_crss}
\alias{predict.baseline_crss}
\title{Predict method for a \code{baseline_crss} object}
\usage{
\method{predict}{baseline_crss}(object, newX, ...)
}
\arguments{
\item{object}{A \code{baseline_crss} object.}

\item{newX}{A numeric matrix at which predictions should be performed}

\item{...}{Additional parameters to be passed down}
}
\value{
A matrix of risk predictions
}
\description{
Predict method for a \code{baseline_crss} object
}
