% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-risk.R
\name{predict.baseline_survival}
\alias{predict.baseline_survival}
\title{Predict method for a \code{baseline_survival} object}
\usage{
\method{predict}{baseline_survival}(object, newX, time_eval, ...)
}
\arguments{
\item{object}{A \code{baseline_survival} object.}

\item{newX}{A numeric matrix at which predictions should be performed.}

\item{time_eval}{The time at which baseline risk shall be predicted. Must be nonnegative numeric vector of length one.}

\item{...}{Additional parameters to be passed down}
}
\value{
A matrix of risk predictions
}
\description{
Predict method for a \code{baseline_survival} object
}
