% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_model_predict.R
\name{predict.effect_model}
\alias{predict.effect_model}
\title{Predict method for a \code{effect_model_crss} object}
\usage{
\method{predict}{effect_model}(object, neww, newX, ...)
}
\arguments{
\item{object}{A \code{effect_model_crss} object.}

\item{neww}{Treatment assignment}

\item{newX}{A numeric matrix at which predictions should be performed}

\item{...}{Additional parameters to be passed down}
}
\value{
A matrix of risk predictions
}
\description{
Predict method for a \code{effect_model_crss} object
}
