% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-modeling.R
\name{predict.risk_model_survival}
\alias{predict.risk_model_survival}
\title{Predict method for a \code{predmod_surv} object}
\usage{
\method{predict}{risk_model_survival}(object, neww, time_eval, newz = NULL, newX = NULL, ...)
}
\arguments{
\item{object}{A \code{predmod_surv} object.}

\item{neww}{Treatment assignment}

\item{time_eval}{The time at which baseline risk shall be predicted. Must be nonnegative numeric vector of length one.}

\item{newz}{TODO}

\item{newX}{A numeric matrix at which predictions should be performed}

\item{...}{Additional parameters to be passed down}
}
\value{
A matrix of risk predictions
}
\description{
Predict method for a \code{predmod_surv} object
}
