% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{predicted.benefit.inference}
\alias{predicted.benefit.inference}
\title{performs inference on the per-individual predicted benefits}
\usage{
predicted.benefit.inference(predicted.benefits, significance.level = 0.05)
}
\arguments{
\item{predicted.benefits}{a vector of per-individual predicted benefits (either absolute of relative)}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
performs inference on the per-individual predicted benefits
}
