% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{predicted_benefit_inference}
\alias{predicted_benefit_inference}
\title{performs inference on the per-individual predicted benefits}
\usage{
predicted_benefit_inference(
  predmod,
  subset = NULL,
  relative = FALSE,
  significance_level = 0.05
)
}
\arguments{
\item{predmod}{a predmod object}

\item{subset}{The indices of the subgroup of interest}

\item{relative}{Relative effect?}

\item{significance_level}{the significance level. Default is 0.05.}
}
\description{
performs inference on the per-individual predicted benefits
}
