% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-modeling-surv.R
\name{prep_predict}
\alias{prep_predict}
\title{creates helper object for predictions in Fine-Gray models}
\usage{
prep_predict(time, status, k = 1)
}
\arguments{
\item{time}{time at risk}

\item{status}{Status. Zero if survived}

\item{k}{Status of interest. Default is 1.}
}
\value{
A list
}
\description{
creates helper object for predictions in Fine-Gray models
}
