% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-variable-at-a-time.R
\name{rate.ratio}
\alias{rate.ratio}
\title{Returns a rateratio object as in the package rateratio.test as well as an estimate of the rate ratio.}
\usage{
rate.ratio(y, w, lifeyears, prediction.timeframe = NULL, subgroup = NULL, ...)
}
\arguments{
\item{y}{A binary vector of outcomes}

\item{w}{A binary vector of treatment status (1 = treatment group)}

\item{lifeyears}{A vector of life-years}

\item{prediction.timeframe}{TODO}

\item{subgroup}{A logical vector indicating a subgroup}

\item{...}{Additional arguments for rateratio.test()}
}
\value{
a rateratio object and an estimate of the rate ratio
}
\description{
Returns a rateratio object as in the package rateratio.test as well as an estimate of the rate ratio.
}
