% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-variable-at-a-time.R
\name{rate_ratio}
\alias{rate_ratio}
\title{Returns a rateratio object as in the package rateratio.test as well as an estimate of the rate ratio.}
\usage{
rate_ratio(status, time, w, subgroup = NULL, ...)
}
\arguments{
\item{status}{Numeric vector with a unique code for each failure type. Code 0 denotes a censored observation.}

\item{time}{vector of failure/censoring times.}

\item{w}{Binary vector of treatment assignment status. Equal to 1 for treatment group and 0 for control group.}

\item{subgroup}{A integer-valued vector indicating a subgroup}

\item{...}{Additional arguments for rateratio.test()}
}
\value{
a rateratio object and an estimate of the rate ratio
}
\description{
Returns a rateratio object as in the package rateratio.test as well as an estimate of the rate ratio.
}
