% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{relative.observed.benefit}
\alias{relative.observed.benefit}
\title{calculates the relative observed benefit, which corresponds to a risk ratio, and the associated confidence interval}
\usage{
relative.observed.benefit(y, w, significance.level = 0.05)
}
\arguments{
\item{y}{vector of binary outcomes}

\item{w}{vector of binary treatment status}

\item{significance.level}{the significance level. Default is 0.05.}
}
\description{
calculates the relative observed benefit, which corresponds to a risk ratio, and the associated confidence interval
}
