% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-modeling.R
\name{risk.model.stage2}
\alias{risk.model.stage2}
\title{perform risk modeling: second stage
\code{y = g(beta_0 +beta_1 * w + beta_2 * z + beta_2 w * z)}}
\usage{
risk.model.stage2(z, y, w, constant.treatment.effect = FALSE, intercept = TRUE)
}
\arguments{
\item{z}{the \code{z} in the 2nd stage, which is used as product with w and as a single regressor.}

\item{y}{a binary response vector}

\item{w}{a binary treatment assignment vector}

\item{constant.treatment.effect}{If TRUE, the interaction z*w is not used as regressor. Default is FALSE.}

\item{intercept}{logical. Shall an intercept be included? Default is \code{FALSE}}
}
\description{
perform risk modeling: second stage
\code{y = g(beta_0 +beta_1 * w + beta_2 * z + beta_2 w * z)}
}
