% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-modeling.R
\name{risk.modeling}
\alias{risk.modeling}
\title{performs risk modeling by penalized logistic regression}
\usage{
risk.modeling(
  X,
  y,
  w,
  alpha = 1,
  z = NULL,
  lifeyears = NULL,
  prediction.timeframe = NULL,
  intercept.stage.2 = TRUE,
  constant.treatment.effect = FALSE
)
}
\arguments{
\item{X}{design matrix or data frame}

\item{y}{a binary response vector}

\item{w}{a binary treatment assignment vector}

\item{alpha}{the alpha as in glmnet. Default is 1 (= Lasso)}

\item{z}{the \code{z} in the 2nd stage, which is used as product with w and as a single regressor. If NULL, thelinear predictor of stage 1 will be used.}

\item{lifeyears}{vector of life years. Default is NULL.}

\item{prediction.timeframe}{vector of the prediction time frame. Default is NULL.}

\item{intercept.stage.2}{logical. Shall an intercept in stage 2 be included? Default is \code{TRUE}}

\item{constant.treatment.effect}{If TRUE, the interaction z*w is not used as regressor in stage 2. Default is FALSE.}
}
\description{
performs risk modeling by penalized logistic regression
}
