% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{subgroup.plot}
\alias{subgroup.plot}
\title{make a subgroup plot of a variable x}
\usage{
subgroup.plot(
  pred.model.obj,
  x,
  relative = FALSE,
  group.bounds = NULL,
  quantile.bounds = c(0.25, 0.5, 0.75),
  quantile.nam = TRUE,
  risk.quantile.bounds = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{pred.model.obj}{prediction model object, as returned by risk.modeling(), effect.modeling(), or grf.modeling()}

\item{x}{a vector that shall be visualized}

\item{relative}{TODO}

\item{group.bounds}{TODO}

\item{quantile.bounds}{TODO}

\item{quantile.nam}{TODO}

\item{risk.quantile.bounds}{TODO}
}
\description{
make a subgroup plot of a variable x
}
