% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{average_treatment_effect}
\alias{average_treatment_effect}
\title{computes the average treatment effect (ATE) of a predmod object}
\usage{
average_treatment_effect(
  x,
  subset = NULL,
  relative = FALSE,
  time_eval = NULL,
  neww = NULL,
  newX = NULL,
  newz = NULL
)
}
\arguments{
\item{x}{A predmod object}

\item{subset}{The indices of the subgroup of interest}

\item{relative}{Shall relative ATE be calculated?}

\item{time_eval}{Time at which to evaluate the failure risk predictions.}
}
\description{
computes the average treatment effect (ATE) of a predmod object
}
