% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{C_benefit}
\alias{C_benefit}
\title{calculates C index for benefit}
\usage{
C_benefit(y, w, pred_ben)
}
\arguments{
\item{y}{binary response vector}

\item{w}{binary treatment assignment vector}

\item{pred_ben}{vector of predicted benefits}
}
\value{
point estimate and standard error of C index for benefit
}
\description{
calculates C index for benefit
}
