% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concordance}
\alias{concordance}
\title{Function to predict concordance}
\usage{
concordance(
  x,
  baseline_risk = NULL,
  risk = NULL,
  benefit = NULL,
  status = NULL,
  w = NULL
)
}
\arguments{
\item{x}{A predmod object of a risk or effect model}

\item{baseline_risk}{the baseline risk estimated by the modle}

\item{risk}{the final risk estimate of the model}

\item{benefit}{vector of estimated benefits}

\item{status}{vector of binary outcomes}

\item{w}{a binary vector of treatment status}
}
\description{
Function to predict concordance
}
