% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{Performs DML estimation by using an interactive regression model}
\usage{
dml(
  X,
  w,
  status,
  ml_g = "random.forest",
  ml_m = "random.forest",
  significance.level = 0.05,
  n_folds = 5,
  score = "ATE",
  trimming_rule = "truncate",
  trimming_threshold = 1e-12,
  dml_procedure = "dml2",
  draw_sample_splitting = TRUE,
  apply_cross_fitting = TRUE
)
}
\arguments{
\item{X}{a matrix or data frame of covariates}

\item{w}{a binary vector of treatment status}

\item{status}{a binary vector of outcomes}

\item{ml_g}{a regression machine learner; refers to the nuisance function \code{g0(X) = E[Y|X,W]}. Either  'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_g = mlr3::lrn("regr.ranger", num.trees = 500, mtry = NULL, min.node.size = NULL, max.depth = NULL) for a classification forest, which is also the default.}

\item{ml_m}{a classification machine learner; refers to the nuisance function \code{m0(X) = E[W|X]}. Either  'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_m = mlr3::lrn("classif.ranger", num.trees = 500, mtry = NULL, min.node.size = NULL, max.depth = NULL) for a classification forest, which is also the default.}

\item{significance.level}{the significance level. Default is 0.05.}

\item{n_folds}{number of folds. Default is 5.}

\item{score}{the score. Default is "ATE"}

\item{trimming_rule}{the trimming rule. Default is "truncate"}

\item{trimming_threshold}{the trimming rule. Default is "1e-12"}

\item{dml_procedure}{the DML procedure. Default is "dm12"}

\item{draw_sample_splitting}{Indicates whether the sample splitting should be drawn during initialization of the object. Default is TRUE.}

\item{apply_cross_fitting}{ndicates whether cross-fitting should be applied. Default is TRUE.}
}
\value{
Estimates of the causal parameter and a 'DoubleML' object
}
\description{
Performs DML using the DoubleML package.
}
\references{
Bach P, Chernozhukov V, Kurz M, Spindler M (2021). “DoubleML – An Object-Oriented Implementation of Double Machine Learning in R.” arXiv: https://arxiv.org/abs/2103.09603, 2103.09603.
}
