% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{get_benefits}
\alias{get_benefits}
\title{Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio}
\usage{
get_benefits(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  baseline_risk = NULL,
  time_eval = NULL,
  odds_ratio = FALSE,
  significance_level = 0.05,
  newX = NULL,
  newstatus = NULL,
  neww = NULL,
  newz = NULL
)
}
\arguments{
\item{x}{prediction model object}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{baseline_risk}{The baseline risk that shall be used for grouping. If \code{NULL} (default), then the baseline risk in \code{x} is used.}

\item{time_eval}{Time at which we evaluate the risk predictions.}

\item{odds_ratio}{Logical. If \code{TRUE}, odds ratios per quantile group will be computed. Default is \code{FALSE}.}

\item{significance_level}{the significance level. Default is 0.05.}

\item{X}{Optional covariate matrix to calculate benefits with}

\item{status}{Optional target variables to calculate benefits with}

\item{w}{Optional treatment assignment variables to calculate benefits with}
}
\description{
Calculates group-level benefits from a prediction model, and the associated confidence intervals.
The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio
}
