% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits-impaccount.R
\name{get_benefits_imputation}
\alias{get_benefits_imputation}
\title{Calculates imputation-adjusted group-level benefits}
\usage{
get_benefits_imputation(
  x,
  cutoffs = c(0.25, 0.5, 0.75),
  breaks = NULL,
  baseline_risk = NULL,
  significance_level = 0.05,
  newX = NULL,
  newstatus = NULL,
  neww = NULL,
  newz = NULL,
  shrunk = FALSE
)
}
\arguments{
\item{x}{a list of prediction model object}

\item{cutoffs}{the quantile cutoff points. Default is c(0.25, 0.5, 0.75), which yields the quartiles.}

\item{breaks}{List of breaks along which to perform the grouping. If passed, overrules the grouping implied by \code{cutoffs}}

\item{baseline_risk}{A list of baseline risks that shall be used for grouping. If \code{NULL} (default), then the baseline risks in \code{x} are used.}

\item{significance_level}{the significance level. Default is 0.05.}

\item{newX}{Optional covariate matrix to calculate benefits with}

\item{newstatus}{Optional list of target variables to calculate benefits with}

\item{neww}{Optional list of treatment assignment variables to calculate benefits with}

\item{newz}{Optional linear predictors}

\item{shrunk}{TODO}
}
\description{
Calculates imputation-adjusted group-level benefits  from a prediction model, and the associated confidence intervals.The returned benefits are the observed and predicted relative and absolute benefits as well as the odds ratio.
}
