% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impaccount_effect_model}
\alias{impaccount_effect_model}
\title{Account for imputation uncertainty in effect models}
\usage{
impaccount_effect_model(x)
}
\arguments{
\item{x}{a list of effect model objects (imputed)}
}
\value{
A list of imputation-accounted stuff
}
\description{
Account for imputation uncertainty in effect models
}
