% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-model-impaccount.R
\name{impaccount_risk_model}
\alias{impaccount_risk_model}
\title{Account for imputation uncertainty in risk models}
\usage{
impaccount_risk_model(x)
}
\arguments{
\item{x}{a list of effect model objects (imputed)}
}
\value{
A list of imputation-accounted absolute and relative benefits, coefficients and risk-predictions
}
\description{
Account for imputation uncertainty in risk models
}
