% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefits.R
\name{predicted_benefit_inference}
\alias{predicted_benefit_inference}
\title{performs inference on the per-individual predicted benefits}
\usage{
predicted_benefit_inference(
  x,
  subset = NULL,
  relative = FALSE,
  significance_level = 0.05,
  neww = NULL,
  newX = NULL,
  newz = NULL,
  shrunk = FALSE
)
}
\arguments{
\item{x}{A predmod object}

\item{subset}{The indices of the subgroup of interest}

\item{relative}{Shall relative ATE be calculated?}

\item{significance_level}{The significance level, default is 0.05}
}
\description{
performs inference on the per-individual predicted benefits
}
