% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-variable-at-a-time.R
\name{rate_ratio}
\alias{rate_ratio}
\title{Evaluate the rate ratio between two groups}
\usage{
rate_ratio(status, time, w, subgroup = NULL, ...)
}
\arguments{
\item{status}{Numeric vector with a unique code for each failure type. Code 0 denotes a censored observation.}

\item{time}{vector of failure/censoring times.}

\item{w}{Binary vector of treatment assignment status. Equal to 1 for treatment group and 0 for control group.}

\item{subgroup}{A integer-valued vector indicating a subgroup}

\item{...}{Additional arguments for rateratio.test()}
}
\value{
a rateratio object containing an estimate of the rate ratio between the two groups, as well as the 95\% confidence interval and p-value
}
\description{
Evaluates the rate ratio between two groups using the rateratio.test package
}
\references{
Fay, M. P. (2010). Two-sided exact tests and matching confidence intervals for discrete data. R
Journal, 2(1), 53-58. DOI: https://doi.org/10.32614/RJ-2010-008
}
